Populate valid versions from metadata, too
Currently, WPScan populates its list of valid WordPress versions
from a list of fingerprinted WordPress files.

However, this file is not always up-to-date, causing version
detection to fail as WPScan thinks its not a valid version
and raises a InvalidWordPressVersion exception.

Since a lot of the version information is also pulled from
metadata.json anyway, also include this file for the initial
version detection, instead of just relying on the list of
fingerprinted files.